[ao] moving private quantize.py fns to utils.py
torch.ao.quantization.quantize can either be a function or a module (torch/ao/quantization/quantize.py) which makes running commands like torch.ao.quantization.quantize.<fn> ambiguous and cause errors in some cases. For this reason we move the private functions of quantize.py to utils.py so that users can use either torch.ao.quantization.utils.<private_fn> or torch.ao.quantization.<public_fn>

Differential Revision: [D41889268](https://our.internmc.facebook.com/intern/diff/D41889268/)

**NOTE FOR REVIEWERS**: This PR has internal Meta-specific changes or comments, please review them on [Phabricator](https://our.internmc.facebook.com/intern/diff/D41889268/)!

[ghstack-poisoned]
