+++ torch/quantization/quantize.py
@@ -8,15 +8,9 @@ here.
 """
 
 from torch.ao.quantization.quantize import _convert
-from torch.ao.quantization.quantize import _observer_forward_hook
 from torch.ao.quantization.quantize import _propagate_qconfig_helper
-from torch.ao.quantization.quantize import _remove_activation_post_process
-from torch.ao.quantization.quantize import _remove_qconfig
-from torch.ao.quantization.quantize import _add_observer_
 from torch.ao.quantization.quantize import add_quant_dequant
 from torch.ao.quantization.quantize import convert
-from torch.ao.quantization.quantize import _get_observer_dict
-from torch.ao.quantization.quantize import _get_unique_devices_
 from torch.ao.quantization.quantize import is_activation_post_process
 from torch.ao.quantization.quantize import prepare
 from torch.ao.quantization.quantize import prepare_qat
@@ -24,5 +18,11 @@ from torch.ao.quantization.quantize import propagate_qconfig_
 from torch.ao.quantization.quantize import quantize
 from torch.ao.quantization.quantize import quantize_dynamic
 from torch.ao.quantization.quantize import quantize_qat
-from torch.ao.quantization.quantize import _register_activation_post_process_hook
 from torch.ao.quantization.quantize import swap_module
+from torch.ao.quantization.utils import _remove_activation_post_process
+from torch.ao.quantization.utils import _remove_qconfig
+from torch.ao.quantization.utils import _add_observer_
+from torch.ao.quantization.utils import _get_observer_dict
+from torch.ao.quantization.utils import _get_unique_devices_
+from torch.ao.quantization.utils import _observer_forward_hook
+from torch.ao.quantization.utils import _register_activation_post_process_hook
